+++ src/main/java/mod/acgaming/universaltweaks/mods/mekanism/dupes/UTMekanismFixes.java
@@ -1,4 +1,4 @@
-package mod.acgaming.universaltweaks.mods.mekanism;
+package mod.acgaming.universaltweaks.mods.mekanism.dupes;
 
 import java.util.ArrayList;
