+++ src/main/resources/mixins.mods.mekanism.fluidtank.json
@@ -1,0 +1,7 @@
+{
+  "package": "mod.acgaming.universaltweaks.mods.mekanism.fluidtank.mixin",
+  "refmap": "universaltweaks.refmap.json",
+  "minVersion": "0.8",
+  "compatibilityLevel": "JAVA_8",
+  "mixins": ["UTTileEntityFluidTankMixin"]
+}
